Exercises 4.15 to 4.23

+++ Part4/blogslist/controllers/blogs.js
@@ -1,11 +1,14 @@
 const blogs = require('express').Router();
 const Blog = require('../models/blogs');
+const User = require('../models/users');
 const { ObjectId } = require('mongodb');
 require('express-async-errors');
+
 
 blogs.get('/', (request, response) => {
   Blog
     .find({})
+    .populate('user', {username: 1, name: 1})
     .then(blogs => {
       response.json(blogs);
     });
@@ -20,16 +23,24 @@ blogs.get('/:id', (request, response) => {
 });
 
 blogs.post('/', async (request, response) => {
+  const body = request.body;
 
-  if (!request.body.title || !request.body.url) {
+  // const token = jwt.verify(request.token, process.env.SECRET);
+  const token = request.token;
+  
+  if (!token) {
+    return response.status(400).send({error: 'Invalid Token'});
+  }
+  const user = await User.findById(token.id);
+  if (!body.title || !body.url) {
     response.status(400).send({ error: 'Bad request' });
     return;
   }
 
-  if (!request.body.likes) {
+  if (!body.likes) {
     request.body.likes = 0;
   }
-  const blog = new Blog(request.body);
+  const blog = new Blog({ ...body, user: user._id});
   
   const result = await blog.save();
   response.status(201).json(result);
@@ -42,17 +53,27 @@ blogs.put('/:id', async (request, response) => {
     title: body.title,
     author: body.author,
     url: body.url,
-    likes: body.likes
+    likes: body.likes,
   };
 
   const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, newBlog, { new: true });
   response.json(updatedBlog);
 });
   
-blogs.delete('/:id', (request, response, next) => {
-  Blog.deleteOne({ _id: new ObjectId(request.params.id) })
-    .then(response.status(204).end())
-    .catch(error => next(error));
+blogs.delete('/:id', async (request, response) => {
+  const token = request.token;
+  if (!token) {
+    return response.status(400).send({error: 'Invalid Token'});
+  }
+
+  const RequestedBlog = await Blog.findById(request.params.id);
+  
+  if (!(token.id === RequestedBlog.user.toString())) {
+    return response.status(405).send({error: 'Operation not allowed'});
+  }
+
+  await Blog.deleteOne({ _id: new ObjectId(request.params.id) });
+  response.status(204).end();
 });
 
 module.exports = blogs;
